// src/generated/graphql.tsx
export type Maybe<T> = T | null
// ...
export type Scalars = {
  ID: string
  String: string
  Boolean: boolean
  Int: number
  Float: number
  /** A date string, such as 2007-12-03 (YYYY-MM-DD), compliant with ISO 8601 standard for representation of dates using the Gregorian calendar. */
  Date: any
  /** A date-time string at UTC, such as 2007-12-03T10:15:30Z, compliant with the date-timeformat outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representationof dates and times using the Gregorian calendar. */
  DateTime: any
  Hex: any
  /** Raw JSON value */
  Json: any
  /** The Long scalar type represents non-fractional signed whole numeric values. Long can represent values between -(2^63) and 2^63 - 1. */
  Long: any
  RGBAHue: any
  RGBATransparency: any
  /** Slate-compatible RichText AST */
  RichTextAST: any
}

export type Aggregate = {
  __typename?: "Aggregate"
  count: Scalars["Int"]
}

/** Asset system model */
export type Asset = Node & {
  __typename?: "Asset"
  avatarAuthor: Array<Author>
  coverPhotoPost: Array<Post>
  /** The time the document was created */
  createdAt: Scalars["DateTime"]
  /** User that created this document */
  createdBy?: Maybe<User>
  /** Get the document in other stages */
  documentInStages: Array<Asset>
  /** The file name */
  fileName: Scalars["String"]
  /** The file handle */
  handle: Scalars["String"]
  /** The height of the file */
  height?: Maybe<Scalars["Float"]>
  /** List of Asset versions */
  history: Array<Version>
  /** The unique identifier */
  id: Scalars["ID"]
  /** System Locale field */
  locale: Locale
  /** Get the other localizations for this document */
  localizations: Array<Asset>
  /** The mime type of the file */
  mimeType?: Maybe<Scalars["String"]>
  /** The time the document was published. Null on documents in draft stage. */
  publishedAt?: Maybe<Scalars["DateTime"]>
  /** User that last published this document */
  publishedBy?: Maybe<User>
  scheduledIn: Array<ScheduledOperation>
  /** The file size */
  size?: Maybe<Scalars["Float"]>
  /** System stage field */
  stage: Stage
  /** The time the document was updated */
  updatedAt: Scalars["DateTime"]
  /** User that last updated this document */
  updatedBy?: Maybe<User>
  /** Get the url for the asset with provided transformations applied. */
  url: Scalars["String"]
  /** The file width */
  width?: Maybe<Scalars["Float"]>
}
// ...
export type AssetConnection = {
  __typename?: "AssetConnection"
  aggregate: Aggregate
  /** A list of edges. */
  edges: Array<AssetEdge>
  /** Information to aid in pagination. */
  pageInfo: PageInfo
}
// ...
export type Author = Node & {
  __typename?: "Author"
  avatar: Asset
  /** The time the document was created */
  createdAt: Scalars["DateTime"]
  /** User that created this document */
  createdBy?: Maybe<User>
  /** Get the document in other stages */
  documentInStages: Array<Author>
  /** List of Author versions */
  history: Array<Version>
  /** The unique identifier */
  id: Scalars["ID"]
  name: Scalars["String"]
  posts: Array<Post>
  /** The time the document was published. Null on documents in draft stage. */
  publishedAt?: Maybe<Scalars["DateTime"]>
  /** User that last published this document */
  publishedBy?: Maybe<User>
  scheduledIn: Array<ScheduledOperation>
  /** System stage field */
  stage: Stage
  /** The time the document was updated */
  updatedAt: Scalars["DateTime"]
  /** User that last updated this document */
  updatedBy?: Maybe<User>
}
// ...
export type AuthorConnection = {
  __typename?: "AuthorConnection"
  aggregate: Aggregate
  /** A list of edges. */
  edges: Array<AuthorEdge>
  /** Information to aid in pagination. */
  pageInfo: PageInfo
}
// ...
export type DocumentVersion = {
  __typename?: "DocumentVersion"
  createdAt: Scalars["DateTime"]
  data?: Maybe<Scalars["Json"]>
  id: Scalars["ID"]
  revision: Scalars["Int"]
  stage: Stage
}
// ...
export enum Locale {
  /** System locale */
  En = "en",
}
// ...
export type Node = {
  /** The id of the object. */
  id: Scalars["ID"]
  /** The Stage of an object */
  stage: Stage
}
// ...
export type Post = Node & {
  __typename?: "Post"
  author?: Maybe<Author>
  content: RichText
  coverPhoto: Asset
  /** The time the document was created */
  createdAt: Scalars["DateTime"]
  /** User that created this document */
  createdBy?: Maybe<User>
  /** Get the document in other stages */
  documentInStages: Array<Post>
  /** List of Post versions */
  history: Array<Version>
  /** The unique identifier */
  id: Scalars["ID"]
  published: Scalars["Date"]
  /** The time the document was published. Null on documents in draft stage. */
  publishedAt?: Maybe<Scalars["DateTime"]>
  /** User that last published this document */
  publishedBy?: Maybe<User>
  scheduledIn: Array<ScheduledOperation>
  slug: Scalars["String"]
  /** System stage field */
  stage: Stage
  title: Scalars["String"]
  /** The time the document was updated */
  updatedAt: Scalars["DateTime"]
  /** User that last updated this document */
  updatedBy?: Maybe<User>
}
// ...
export type PostConnection = {
  __typename?: "PostConnection"
  aggregate: Aggregate
  /** A list of edges. */
  edges: Array<PostEdge>
  /** Information to aid in pagination. */
  pageInfo: PageInfo
}
// ...
export type Query = {
  __typename?: "Query"
  /** Retrieve a single asset */
  asset?: Maybe<Asset>
  /** Retrieve document version */
  assetVersion?: Maybe<DocumentVersion>
  /** Retrieve multiple assets */
  assets: Array<Asset>
  /** Retrieve multiple assets using the Relay connection interface */
  assetsConnection: AssetConnection
  /** Retrieve a single author */
  author?: Maybe<Author>
  /** Retrieve document version */
  authorVersion?: Maybe<DocumentVersion>
  /** Retrieve multiple authors */
  authors: Array<Author>
  /** Retrieve multiple authors using the Relay connection interface */
  authorsConnection: AuthorConnection
  /** Fetches an object given its ID */
  node?: Maybe<Node>
  /** Retrieve a single post */
  post?: Maybe<Post>
  /** Retrieve document version */
  postVersion?: Maybe<DocumentVersion>
  /** Retrieve multiple posts */
  posts: Array<Post>
  /** Retrieve multiple posts using the Relay connection interface */
  postsConnection: PostConnection
  /** Retrieve a single scheduledOperation */
  scheduledOperation?: Maybe<ScheduledOperation>
  /** Retrieve multiple scheduledOperations */
  scheduledOperations: Array<ScheduledOperation>
  /** Retrieve multiple scheduledOperations using the Relay connection interface */
  scheduledOperationsConnection: ScheduledOperationConnection
  /** Retrieve a single scheduledRelease */
  scheduledRelease?: Maybe<ScheduledRelease>
  /** Retrieve multiple scheduledReleases */
  scheduledReleases: Array<ScheduledRelease>
  /** Retrieve multiple scheduledReleases using the Relay connection interface */
  scheduledReleasesConnection: ScheduledReleaseConnection
  /** Retrieve a single user */
  user?: Maybe<User>
  /** Retrieve multiple users */
  users: Array<User>
  /** Retrieve multiple users using the Relay connection interface */
  usersConnection: UserConnection
}
// ...
export type RichText = {
  __typename?: "RichText"
  /** Returns HTMl representation */
  html: Scalars["String"]
  /** Returns Markdown representation */
  markdown: Scalars["String"]
  /** Returns AST representation */
  raw: Scalars["RichTextAST"]
  /** Returns plain-text contents of RichText */
  text: Scalars["String"]
}

/** Scheduled Operation system model */
export type ScheduledOperation = Node & {
  __typename?: "ScheduledOperation"
  affectedDocuments: Array<ScheduledOperationAffectedDocument>
  /** The time the document was created */
  createdAt: Scalars["DateTime"]
  /** User that created this document */
  createdBy?: Maybe<User>
  /** Operation description */
  description?: Maybe<Scalars["String"]>
  /** Get the document in other stages */
  documentInStages: Array<ScheduledOperation>
  /** Operation error message */
  errorMessage?: Maybe<Scalars["String"]>
  /** The unique identifier */
  id: Scalars["ID"]
  /** The time the document was published. Null on documents in draft stage. */
  publishedAt?: Maybe<Scalars["DateTime"]>
  /** User that last published this document */
  publishedBy?: Maybe<User>
  /** Raw operation payload including all details, this field is subject to change */
  rawPayload: Scalars["Json"]
  /** The release this operation is scheduled for */
  release?: Maybe<ScheduledRelease>
  /** System stage field */
  stage: Stage
  /** operation Status */
  status: ScheduledOperationStatus
  /** The time the document was updated */
  updatedAt: Scalars["DateTime"]
  /** User that last updated this document */
  updatedBy?: Maybe<User>
}
// ...
export type ScheduledOperationConnection = {
  __typename?: "ScheduledOperationConnection"
  aggregate: Aggregate
  /** A list of edges. */
  edges: Array<ScheduledOperationEdge>
  /** Information to aid in pagination. */
  pageInfo: PageInfo
}
// ...
export type ScheduledRelease = Node & {
  __typename?: "ScheduledRelease"
  /** The time the document was created */
  createdAt: Scalars["DateTime"]
  /** User that created this document */
  createdBy?: Maybe<User>
  /** Release description */
  description?: Maybe<Scalars["String"]>
  /** Get the document in other stages */
  documentInStages: Array<ScheduledRelease>
  /** Release error message */
  errorMessage?: Maybe<Scalars["String"]>
  /** The unique identifier */
  id: Scalars["ID"]
  /** Whether scheduled release should be run */
  isActive: Scalars["Boolean"]
  /** Whether scheduled release is implicit */
  isImplicit: Scalars["Boolean"]
  /** Operations to run with this release */
  operations: Array<ScheduledOperation>
  /** The time the document was published. Null on documents in draft stage. */
  publishedAt?: Maybe<Scalars["DateTime"]>
  /** User that last published this document */
  publishedBy?: Maybe<User>
  /** Release date and time */
  releaseAt?: Maybe<Scalars["DateTime"]>
  /** System stage field */
  stage: Stage
  /** Release Status */
  status: ScheduledReleaseStatus
  /** Release Title */
  title?: Maybe<Scalars["String"]>
  /** The time the document was updated */
  updatedAt: Scalars["DateTime"]
  /** User that last updated this document */
  updatedBy?: Maybe<User>
}
// ...
export type ScheduledReleaseConnection = {
  __typename?: "ScheduledReleaseConnection"
  aggregate: Aggregate
  /** A list of edges. */
  edges: Array<ScheduledReleaseEdge>
  /** Information to aid in pagination. */
  pageInfo: PageInfo
}
// ...
export enum Stage {
  /** The Draft is the default stage for all your content. */
  Draft = "DRAFT",
  /** The Published stage is where you can publish your content to. */
  Published = "PUBLISHED",
}

export enum SystemDateTimeFieldVariation {
  Base = "BASE",
  Combined = "COMBINED",
  Localization = "LOCALIZATION",
}

export type UnpublishLocaleInput = {
  /** Locales to unpublish */
  locale: Locale
  /** Stages to unpublish selected locales from */
  stages: Array<Stage>
}

/** User system model */
export type User = Node & {
  __typename?: "User"
  /** The time the document was created */
  createdAt: Scalars["DateTime"]
  /** Get the document in other stages */
  documentInStages: Array<User>
  /** The unique identifier */
  id: Scalars["ID"]
  /** Flag to determine if user is active or not */
  isActive: Scalars["Boolean"]
  /** User Kind. Can be either MEMBER, PAT or PUBLIC */
  kind: UserKind
  /** The username */
  name: Scalars["String"]
  /** Profile Picture url */
  picture?: Maybe<Scalars["String"]>
  /** The time the document was published. Null on documents in draft stage. */
  publishedAt?: Maybe<Scalars["DateTime"]>
  /** System stage field */
  stage: Stage
  /** The time the document was updated */
  updatedAt: Scalars["DateTime"]
}
// ...
export type UserConnection = {
  __typename?: "UserConnection"
  aggregate: Aggregate
  /** A list of edges. */
  edges: Array<UserEdge>
  /** Information to aid in pagination. */
  pageInfo: PageInfo
}
// ...
export type Version = {
  __typename?: "Version"
  createdAt: Scalars["DateTime"]
  id: Scalars["ID"]
  revision: Scalars["Int"]
  stage: Stage
}
// ...
export type GetPostQuery = {
  __typename?: "Query"
  post?: {
    __typename?: "Post"
    id: string
    title: string
    slug: string
    content: { __typename?: "RichText"; html: string }
    coverPhoto: { __typename?: "Asset"; url: string }
  } | null
}